+++ examples/scripts/schedule.ps1
@@ -1,0 +1,3 @@
+{
+    'Hello, there!' | Out-PodeHost
+}