+++ src/Private/PodeServer.ps1
@@ -93,7 +93,7 @@ function Start-PodeSocketServer
     }
 
     # start the runspace for listening on x-number of threads
-    1..$PodeContext.Threads | ForEach-Object {
+    1..$PodeContext.Threads.Web | ForEach-Object {
         Add-PodeRunspace -Type 'Main' -ScriptBlock $listenScript `
             -Parameters @{ 'ThreadId' = $_ }
     }