+++ src/Private/SignalServer.ps1
@@ -76,7 +76,7 @@ function Start-PodeSignalServer
     }
 
     # start the runspace for listening on x-number of threads
-    1..$PodeContext.Threads | ForEach-Object {
+    1..$PodeContext.Threads.Web | ForEach-Object {
         Add-PodeRunspace -Type 'Signals' -ScriptBlock $listenScript `
             -Parameters @{ 'ThreadId' = $_ }
     }